backend/requirements.txt
# =============================================================================
# EVA-Dev Backend Dependencies
# =============================================================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP Client & WebSocket
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Azure Speech SDK
azure-cognitiveservices-speech==1.34.0

# Anthropic Claude API
anthropic==0.34.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database & Caching
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1

# Audio Processing
pydub==0.25.1
numpy==1.25.2
scipy==1.11.4

# Utilities
aiofiles==23.2.1
python-json-logger==2.0.7
structlog==23.2.0

# Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6

# Monitoring & Health Checks
prometheus-client==0.19.0

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
pytest-mock==3.12.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Production WSGI Server
gunicorn==21.2.0
